# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API document for UAMS backend server",
        "title": "Backend API",
        "contact": {
            "name": "Mr. Khai"
        },
        "version": "1.0"
    },
    "host": "http://iot.hcmue.space:8079",
    "basePath": "/v1",
    "paths": {
...
        "/v1/user_accesses": {
            "get": {
                "description": "find all user access log",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_accesses/area_id/{area_id}": {
            "get": {
                "description": "find all user access log by Area Id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log by Area Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_accesses/area_id/{area_id}/period/{from}/{to}": {
            "get": {
                "description": "find all user access log by Area Id and TimeRange",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log by Area Id and TimeRange",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_accesses/period/{from}/{to}": {
            "get": {
                "description": "find all user access log by TimeRange",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log by TimeRange",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_accesses/user_id/{id}": {
            "get": {
                "description": "find all user access log by user id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log by User ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_accesses/user_id/{id}/area_id/{area_id}": {
            "get": {
                "description": "find all user access log by user id and Area Id",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log by User ID and Area Id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_accesses/user_id/{id}/area_id/{area_id}/period/{from}/{to}": {
            "get": {
                "description": "find all user access log by user id and Area Id and TimeRange",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log by User ID and Area Id and TimeRange",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/user_accesses/user_id/{id}/period/{from}/{to}": {
            "get": {
                "description": "find all user access log by user id and Time Range",
                "produces": [
                    "application/json"
                ],
                "summary": "Find User Access log by User ID and Time Range",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.UserAccess"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.UserAccess": {
            "type": "object",
            "properties": {
                "area_id": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "random": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorMsg": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}